Refactored APICalls
