+++ src/main/react-front-end/src/views/Transfer/BrowseModuleComponent.js
@@ -2,7 +2,8 @@
 
 import React, { Component } from 'react';
 import Button from '@material-ui/core/Button';
-import {openDropboxOAuth, openGoogleDriveOAuth, openGridFtpOAuth, openBoxOAuth, savedCredList} from "../../APICalls/APICalls";
+import { openDropboxOAuth, openGoogleDriveOAuth, openGridFtpOAuth, openBoxOAuth } from "../../APICalls/EndpointAPICalls";
+import { savedCredList } from "../../APICalls/APICalls";
 import {store} from "../../App";
 import PropTypes from "prop-types";
 import {cookies} from "../../model/reducers.js";
@@ -187,12 +188,12 @@ export default class BrowseModuleComponent extends Component {
 					<Icon className={'fas fa-bold'} style={iconStyle}/>
                     Box
                 </Button>
-	      		{/* <Button id={endpoint.side + "GridFTP"} style={buttonStyle} disabled={!gridftpIsOpen} onClick={() =>{
-	      			this.credentialTypeExistsThenDo(GRIDFTP_NAME, loginPrep(GRIDFTP_TYPE), openGridFtpOAuth);
-	      		}}>
-	      			<Icon className={'fas fa-server'} style={iconStyle}/>
-	      		GridFTP
-		      	</Button>  */}
+				{/* <Button id={endpoint.side + "GridFTP"} style={buttonStyle} hidden="true	" disabled={!gridftpIsOpen} onClick={() =>{
+					this.credentialTypeExistsThenDo(GRIDFTP_NAME, loginPrep(GRIDFTP_TYPE), openGridFtpOAuth);
+				}}>
+					<Icon className={'fas fa-server'} style={iconStyle}/>
+				GridFTP
+				</Button>  */}
 				<Button id={endpoint.side + "HTTP"} style={buttonStyle} disabled={oneSideIsLoggedInAsGridftp} onClick={() =>{
 	      			loginPrep(HTTP_TYPE)()
 	      		}}>
